Fix a collection of minor style violations.
This enables several new StyleCopAnalyzer rules to
be enabled immediately during migration.

+++ OpenRA.Mods.Common/SpriteLoaders/ShpTDLoader.cs
@@ -217,7 +217,7 @@ namespace OpenRA.Mods.Common.SpriteLoaders
 				bw.Write((ushort)0);
 				bw.Write((ushort)size.Width);
 				bw.Write((ushort)size.Height);
-				bw.Write((uint)0);
+				bw.Write(0U);
 
 				foreach (var f in compressedFrames)
 				{
